cleanup server setup
